Update MyLibraryTest.java
This is what I last was testing when my eclipse was working. There are some errors in the implementation of my attempts to simulate the System.in input, and I would love any suggestions on better ideas, since this is all that I could figure to do after like a dozen hours of research. Obviously, the actual assertions need to be written, but that will take like 10 minutes, once the System.in part is figured out, it's just a matter of simulating user input that tripped me up.